requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0